make: remove 'buildvcs=false' from manifest gen
This was previously introduced in d3df2278dd03487adba93b31493426fc81459e99,
and was required there to handle git worktrees because it was running in
a container, which might not have had access to the .git directory
outside the worktree.

However, because 9a7a79b93efa34286648c146559e599d7e2ac65e switched back
to running it outside the container, it's no longer required.